requirements.txt
fastapi
uvicorn
langgraph
langchain-openai
python-dotenv
pydantic